# internal/docs/swagger.yaml
definitions:
  dto.SongCreateRequest:
    properties:
      group:
        type: string
      song:
        type: string
    type: object
  model.Song:
    properties:
      group:
        type: string
      id:
        type: integer
      link:
        type: string
      releaseDate:
        type: string
      song:
        type: string
      text:
        type: string
    type: object
info:
  contact: {}
  description: REST API для управления онлайн-библиотекой песен.
  title: Online Music Library API
  version: "1.0"
paths:
  /info:
    get:
      consumes:
      - application/json
      description: Возвращает песни с поддержкой пагинации
      parameters:
      - description: Название группы
        in: query
        name: group
        type: string
      - description: Название песни
        in: query
        name: song
        type: string
      - description: Номер страницы (по умолчанию 1)
        in: query
        name: page
        required: true
        type: integer
      - description: Количество куплетов на странице (по умолчанию 2)
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Song'
            type: array
      summary: Получение песни
      tags:
      - songs
  /songs:
    post:
      consumes:
      - application/json
      description: Добавляет новую песню в библиотеку
      parameters:
      - description: Данные песни
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/dto.SongCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            type: string
      summary: Добавление новой песни
      tags:
      - songs
  /songs/{id}:
    delete:
      description: Удаляет песню по её идентификатору
      parameters:
      - description: Идентификатор песни
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Ok
          schema:
            type: string
      summary: Удаление песни
      tags:
      - songs
  /songs/{id}/lyrics:
    get:
      consumes:
      - application/json
      description: Возвращает текст песни с поддержкой пагинации по куплетам
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      - description: Номер страницы (по умолчанию 1)
        in: query
        name: page
        required: true
        type: integer
      - description: Количество куплетов на странице (по умолчанию 2)
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: Получение текста песни
      tags:
      - songs
swagger: "2.0"